+++ CHANGELOG.md
@@ -12,7 +12,8 @@
 - [IconButton] - from Buttons/IconButton to Actions/IconButton
 - [MultiActionButton] - from Buttons/MultiActionButton to Actions/MultiActionButton
 - [MapControl] - from Buttons/MapControl to Actions/MapControl
-- [SSOButtons] - from presets to Actions/Button/preset/SSOButtons
+- [SSOButtons] - from presets to Forms/Actions/Button/preset/SSOButtons
+- [Menu] - from Overlays/Menu to Forms/Actions/Menu
 
 ## [v2.131.0] - 09-10-2025